chore(*): use a single webpack config
Ref: #54
