+++ applications/apps/auth/node/src/index.ts
@@ -1,0 +1,1 @@
+import "./main";
