+++ applications/apps/logs/node/webpack.config.js
@@ -1,37 +1,16 @@
 const { resolve } = require("path");
+const config = require("../../../webpack.config");
 
-const config = (env) => {
-    const isProduction = env.production;
-    const filename = "node.js";
-    const cfg = {
-        mode: isProduction ? "production" : "development",
-        entry: [
-            "./src/main.ts"
-        ],
-        output: {
-            path: resolve(__dirname, "../dist/node"),
-            filename,
-            clean: true
-        },
-        target: "node",
-        node: {
-            __dirname: false,
-            __filename: false,
-        },
-        module: {
-            rules: [
-                {
-                    test: /\.ts$/,
-                    use: "ts-loader",
-                    exclude: /node_modules/
-                },
-            ]
-        },
-        resolve: {
-            extensions: [".ts", ".js"]
-        }
-    };
-    return cfg;
+module.exports = (env) => {
+  const filename = "index.js";
+  return {
+    ...config(env, __dirname),
+    output: {
+      path: resolve(__dirname, "./dist/node"),
+      filename,
+      clean: true,
+      libraryTarget: "commonjs2"
+    },
+    target: "node"
+  }
 };
-
-module.exports = config;
