+++ applications/packages/core/core.env/webpack.config.js
